% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/betaselectr-package.R
\docType{package}
\name{betaselectr-package}
\alias{betaselectr}
\alias{betaselectr-package}
\title{betaselectr: Betas-Select in Structural Equation Models and Linear Models}
\description{
\if{html}{\figure{logo.png}{options: style='float: right' alt='logo' width='120'}}

It computes betas-select, coefficients after standardization in structural equation models and regression models, standardizing only selected variables. Supports models with moderation, with product terms formed after standardization. It also offers confidence intervals that account for standardization, including bootstrap confidence intervals as proposed by Cheung et al. (2022) \doi{10.1037/hea0001188}. An introduction to the package can be found in Sun et al. (2026) \doi{10.1080/00273171.2026.2672692}.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://sfcheung.github.io/betaselectr/}
  \item Report bugs at \url{https://github.com/sfcheung/betaselectr/issues}
}

}
\author{
\strong{Maintainer}: Shu Fai Cheung \email{shufai.cheung@gmail.com} (\href{https://orcid.org/0000-0002-9871-9448}{ORCID})

Authors:
\itemize{
  \item Shu Fai Cheung \email{shufai.cheung@gmail.com} (\href{https://orcid.org/0000-0002-9871-9448}{ORCID})
  \item Rong Wei Sun (\href{https://orcid.org/0000-0003-0034-1422}{ORCID})
  \item Florbela Chang (\href{https://orcid.org/0009-0003-9931-501X}{ORCID})
  \item Wendie Yang (\href{https://orcid.org/0009-0000-8388-6481}{ORCID})
  \item Sing-Hang Cheung (\href{https://orcid.org/0000-0001-5182-0752}{ORCID})
}

}
\keyword{internal}
