% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lm_betaselect_methods.R
\name{deviance.lm_betaselect}
\alias{deviance.lm_betaselect}
\title{Model Deviance}
\usage{
\method{deviance}{lm_betaselect}(
  object,
  type = c("beta", "standardized", "raw", "unstandardized"),
  ...
)
}
\arguments{
\item{object}{A \code{lm_betaselect}-class
object.}

\item{type}{The model from which the
deviance is returned. For
\code{"beta"} or \code{"standardized"}, the
model is the one after selected
variables standardized. For \code{"raw"}
or \code{"unstandardized"}, the model is
the one before standardization was
done.}

\item{...}{Optional arguments.
To be passed to \code{\link[stats:deviance]{stats::deviance()}}.}
}
\value{
It returns the value of the
deviance of the requested model.
}
\description{
Extract the
deviance from a \code{lm_betaselect}
object.
}
\details{
It simply passes the model with
or without selected variables
standardized to \code{\link[stats:deviance]{stats::deviance()}}.
Please refer to
the help page of \code{\link[stats:deviance]{stats::deviance()}}
for details.
}
\examples{

data(data_test_mod_cat)

lm_beta_x <- lm_betaselect(dv ~ iv*mod + cov1 + cat1,
                           data = data_test_mod_cat,
                           to_standardize = "iv")
deviance(lm_beta_x)
deviance(lm_beta_x, type = "raw")

}
\seealso{
\code{\link[=lm_betaselect]{lm_betaselect()}} and \code{\link[stats:deviance]{stats::deviance()}}
}
\author{
Shu Fai Cheung \url{https://orcid.org/0000-0002-9871-9448}
}
