% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lm_betaselect_methods.R
\name{extractAIC.lm_betaselect}
\alias{extractAIC.lm_betaselect}
\title{Extract AIC}
\usage{
\method{extractAIC}{lm_betaselect}(
  fit,
  scale,
  k = 2,
  type = c("beta", "standardized", "raw", "unstandardized"),
  ...
)
}
\arguments{
\item{fit}{A \code{lm_betaselect}-class
object.}

\item{scale}{To be passed
to \code{\link[stats:extractAIC]{stats::extractAIC()}}. See its
help page for details.}

\item{k}{The weight of the
equivalent degrees of freedom to be
used in the computation of AIC.
See \code{\link[stats:extractAIC]{stats::extractAIC()}} for details.}

\item{type}{The model from which the
AIC is returned. For
\code{"beta"} or \code{"standardized"}, the
model is the one after selected
variables standardized. For \code{"raw"}
or \code{"unstandardized"}, the model is
the one before standardization was
done.}

\item{...}{Optional arguments.
To be passed to \code{\link[stats:extractAIC]{stats::extractAIC()}}.}
}
\value{
It returns a numeric vector of
two elements, which is simply the
output of \code{\link[stats:extractAIC]{stats::extractAIC()}}
on the requested model.
}
\description{
Extract the
Akaike Information Criterion (AIC)
from a \code{lm_betaselect} object.
}
\details{
It simply passes the model with
or without selected variables
standardized to \code{\link[stats:extractAIC]{stats::extractAIC()}}.
Please refer to
the help page of \code{\link[stats:extractAIC]{stats::extractAIC()}}
for details.
}
\examples{

data(data_test_mod_cat)

lm_beta_x <- lm_betaselect(dv ~ iv*mod + cov1 + cat1,
                           data = data_test_mod_cat,
                           to_standardize = "iv")
extractAIC(lm_beta_x)
extractAIC(lm_beta_x, type = "raw")

}
\seealso{
\code{\link[=lm_betaselect]{lm_betaselect()}} and \code{\link[stats:extractAIC]{stats::extractAIC()}}
}
\author{
Shu Fai Cheung \url{https://orcid.org/0000-0002-9871-9448}
}
