% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lm_betaselect_methods.R
\name{logLik.lm_betaselect}
\alias{logLik.lm_betaselect}
\title{Extract Log-Likelihood}
\usage{
\method{logLik}{lm_betaselect}(
  object,
  REML = FALSE,
  type = c("beta", "standardized", "raw", "unstandardized"),
  ...
)
}
\arguments{
\item{object}{A \code{lm_betaselect}-class
object.}

\item{REML}{Whether the restricted
log-likelihood is returned. Default
is \code{FALSE}.}

\item{type}{The model from which the
log-likelihood is returned. For
\code{"beta"} or \code{"standardized"}, the
model is the one after selected
variables standardized. For \code{"raw"}
or \code{"unstandardized"}, the model is
the one before standardization was
done.}

\item{...}{Optional arguments.
To be passed to \code{\link[stats:logLik]{stats::logLik.lm()}}.}
}
\value{
It returns an object of the class
\code{logLik}, the same object returned
by \code{\link[stats:logLik]{stats::logLik.lm()}}.
}
\description{
Extract the
log-likelihood of a \code{lm_betaselect}
object.
}
\details{
It simply passes the model with
or without selected variables
standardized to the method
\code{\link[stats:logLik]{stats::logLik.lm()}}. Please refer to
the help page of \code{\link[stats:logLik]{stats::logLik.lm()}}
for details.
}
\examples{

data(data_test_mod_cat)

lm_beta_x <- lm_betaselect(dv ~ iv*mod + cov1 + cat1,
                           data = data_test_mod_cat,
                           to_standardize = "iv")
logLik(lm_beta_x)
logLik(lm_beta_x, type = "raw")

}
\seealso{
\code{\link[=lm_betaselect]{lm_betaselect()}} and \code{\link[stats:logLik]{stats::logLik.lm()}}
}
\author{
Shu Fai Cheung \url{https://orcid.org/0000-0002-9871-9448}
}
