% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lm_betaselect_methods.R
\name{plot.lm_betaselect}
\alias{plot.lm_betaselect}
\title{Plot Diagnostics for an \code{lm_betaselect} Object}
\usage{
\method{plot}{lm_betaselect}(x, model_type = c("beta", "standardized", "raw", "unstandardized"), ...)
}
\arguments{
\item{x}{A \code{lm_betaselect}-class
object.}

\item{model_type}{The model from which the
the diagnostics are plotted For
\code{"beta"} or \code{"standardized"}, the
model is the one after selected
variables standardized. For \code{"raw"}
or \code{"unstandardized"}, the model is
the one before standardization was
done.}

\item{...}{Arguments
to be passed to \code{\link[stats:plot.lm]{stats::plot.lm()}}.
Please refer to the help page of
\code{\link[stats:plot.lm]{stats::plot.lm()}}.}
}
\value{
It returns \code{NULL}. Called for its
side effects.
}
\description{
Plot the diagnostics
for the model before or after
standardization.
}
\details{
It simply passes the model with
or without selected variables
standardized to the \code{plot} method
of \code{lm} objects.
Please refer to
the help page of \code{\link[stats:plot.lm]{stats::plot.lm()}}
for details.
\subsection{IMPORTANT}{

Some diagnostics that makes use
of the sampling variances and
covariances of coefficient estimates
\emph{may} not be applicable to the
models with one or more variables
standardized. Therefore, they should
only be used for exploratory purpose.
}
}
\examples{

data(data_test_mod_cat)

lm_beta_x <- lm_betaselect(dv ~ iv*mod + cov1 + cat1,
                           data = data_test_mod_cat,
                           to_standardize = "iv")
plot(lm_beta_x)
plot(lm_beta_x, model_type = "raw")

}
\seealso{
\code{\link[=lm_betaselect]{lm_betaselect()}} and \code{\link[stats:plot.lm]{stats::plot.lm()}}
}
\author{
Shu Fai Cheung \url{https://orcid.org/0000-0002-9871-9448}
}
