% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lm_betaselect_methods.R
\name{update.lm_betaselect}
\alias{update.lm_betaselect}
\alias{getCall.lm_betaselect}
\title{Update and Re-fit a Call to
'lm_betaselect()'}
\usage{
\method{update}{lm_betaselect}(object, formula., ..., evaluate = TRUE)

\method{getCall}{lm_betaselect}(
  x,
  what = c("lm_betaselect", "beta", "standardized", "raw", "unstandardized"),
  ...
)
}
\arguments{
\item{object}{An \code{lm_betaselect}-class
object.}

\item{formula.}{Changes to the formula,
as in the default \code{\link[=update]{update()}} method.}

\item{...}{For \code{\link[=update.lm_betaselect]{update.lm_betaselect()}},
additional arguments
to the call, as in the default
\code{\link[=update]{update()}} method. Ignored by
\code{\link[=getCall.lm_betaselect]{getCall.lm_betaselect()}}.}

\item{evaluate}{Whether the updated
call will be evaluated. Default is
\code{TRUE}.}

\item{x}{An \code{lm_betaselect}-class
object from which the call is to
be extracted.}

\item{what}{Which call to extract.
For \code{"lm_betaselect"}, the call
to \code{\link[=lm_betaselect]{lm_betaselect()}} is extracted.
For
\code{"beta"} or \code{"standardized"}, the
call used to fit the model \emph{after}
selected variables standardized
is extracted.
For \code{"raw"}
or \code{"unstandardized"}, the call used
to fit hte model \emph{before} standardization
is extracted.}
}
\value{
It returns the output of
\code{\link[=lm_betaselect]{lm_betaselect()}} with the updated
call, such as the updated model.
}
\description{
The \code{update}-method
for an \code{lm_betaselect}-class objects.
}
\details{
This works in the same way
the default \code{update}-method does for
the output of \code{\link[stats:lm]{stats::lm()}}.
}
\examples{

data(data_test_mod_cat)

lm_beta_x <- lm_betaselect(dv ~ iv*mod + cov1,
                           data = data_test_mod_cat,
                           to_standardize = "iv")
summary(lm_beta_x)
lm_beta_x2 <- update(lm_beta_x, ~ . + cat1)
summary(lm_beta_x2)

}
\seealso{
\code{\link[=lm_betaselect]{lm_betaselect()}} and \code{\link[stats:update]{stats::update()}}
}
\author{
Shu Fai Cheung \url{https://orcid.org/0000-0002-9871-9448}
}
