% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lm_betaselect_methods.R
\name{anova.lm_betaselect}
\alias{anova.lm_betaselect}
\title{ANOVA Tables for a
'lm_betaselect'-Class Object}
\usage{
\method{anova}{lm_betaselect}(object, ..., type = c("beta", "standardized", "raw", "unstandardized"))
}
\arguments{
\item{object}{The output of
\code{\link[=lm_betaselect]{lm_betaselect()}}.}

\item{...}{Additional outputs
of \code{\link[=lm_betaselect]{lm_betaselect()}}.}

\item{type}{String. If
\code{"unstandardized"} or \code{"raw"}, the
output \emph{before} standardization
are used If \code{"beta"} or
\code{"standardized"}, then the
output \emph{after} selected
variables standardized are returned.
Default is \code{"beta"}.}
}
\value{
It returns an object of class
\code{anova}, which is identical to
the output of \code{\link[=stats:anova]{stats:anova()}} in
structure.
}
\description{
Return the analysis
of variance tables for
the outputs of
\code{\link[=lm_betaselect]{lm_betaselect()}}.
}
\details{
By default, it calls \code{\link[stats:anova]{stats::anova()}}
on the results with selected variables
standardized. By setting \code{type} to
\code{"raw"} or \code{"unstandardized"}, it
calls \code{\link[stats:anova]{stats::anova()}} on the results
\emph{before} standardization.
}
\examples{

data(data_test_mod_cat)

lm_beta_x <- lm_betaselect(dv ~ iv*mod + cov1 + cat1,
                           data = data_test_mod_cat,
                           to_standardize = "iv",
                           do_boot = FALSE)
anova(lm_beta_x)
anova(lm_beta_x, type = "raw")

}
\seealso{
\code{\link[=lm_betaselect]{lm_betaselect()}}
}
\author{
Shu Fai Cheung \url{https://orcid.org/0000-0002-9871-9448}
}
