% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lm_betaselect_methods.R
\name{coef.lm_betaselect}
\alias{coef.lm_betaselect}
\title{Coefficients of an
'lm_betaselect'-Class Object}
\usage{
\method{coef}{lm_betaselect}(
  object,
  complete = FALSE,
  type = c("beta", "standardized", "raw", "unstandardized"),
  ...
)
}
\arguments{
\item{object}{The output of
\code{\link[=lm_betaselect]{lm_betaselect()}}, or an
\code{lm_betaselect}-class object.}

\item{complete}{If \code{TRUE}, it returns
the full vector of coefficients,
including those of terms dropped in
an over-determined system. See
\code{\link[stats:coef]{stats::coef()}} for further
information. Default is \code{FALSE}.}

\item{type}{String. If \code{"unstandardized"}
or \code{"raw"}, the coefficients \emph{before}
standardization are returned. If
\code{"beta"} or \code{"standardized"}, then
the coefficients \emph{after} selected
variables standardized are returned.
Default is \code{"beta"}.}

\item{...}{Other arguments. Ignored.}
}
\value{
A scalar vector: The estimate of
regression coefficients.
}
\description{
Return the estimates of
coefficients in an
\code{lm_betaselect}-class object.
}
\details{
By default, it extracts the
regression coefficients \emph{after} the
selected variables have been
standardized. If requested, it can
also return the regression
coefficients \emph{without}
standardization.
}
\examples{

data(data_test_mod_cat)

lm_beta_x <- lm_betaselect(dv ~ iv*mod + cov1 + cat1,
                           data = data_test_mod_cat,
                           to_standardize = "iv")
coef(lm_beta_x)
coef(lm_beta_x, type = "raw")

}
\seealso{
\code{\link[=lm_betaselect]{lm_betaselect()}}
}
\author{
Shu Fai Cheung \url{https://orcid.org/0000-0002-9871-9448}
}
