% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lm_betaselect_methods.R
\name{confint.lm_betaselect}
\alias{confint.lm_betaselect}
\title{Confidence Interval for an
'lm_betaselect'-Class Object}
\usage{
\method{confint}{lm_betaselect}(
  object,
  parm,
  level = 0.95,
  method = c("boot", "bootstrap", "ls"),
  type = c("beta", "standardized", "raw", "unstandardized"),
  warn = TRUE,
  boot_type = c("perc", "bc"),
  ...
)
}
\arguments{
\item{object}{The output of
\code{\link[=lm_betaselect]{lm_betaselect()}}.}

\item{parm}{The terms for which
the confidence intervals are returned.
If missing, the confidence intervals
of all terms will be returned.}

\item{level}{The level of confidence,
default is .95, returning the 95\%
confidence interval.}

\item{method}{The method used to
compute the confidence intervals/
If bootstrapping was
requested when calling
\code{\link[=lm_betaselect]{lm_betaselect()}} and this argument
is set to \code{"bootstrap"} or \code{"boot"},
the bootstrap confidence intervals
are returned. If bootstrapping
was not requested or if this argument
is set to \code{"ls"}, then the usual \code{lm}
confidence intervals are
returned, with a warning raised
unless \code{type} is \code{"raw"} or
\verb{"unstandardized".}
Default is \code{"boot"}.}

\item{type}{String. If
\code{"unstandardized"} or \code{"raw"}, the
confidence intervals of the
coefficients \emph{before} standardization
are returned. If \code{"beta"} or
\code{"standardized"}, then the
confidence intervals of the
coefficients \emph{after} selected
variables standardized are returned.
Default is \code{"beta"}.}

\item{warn}{Logical. Whether a warning
will be raised is OLS (or WLS)
confidence intervals are
requested for the model with some
variables standardized (i.e., \code{type}
is \code{"beta"} or \code{"standardized"}).
Default is \code{TRUE}.}

\item{boot_type}{The type of
bootstrap confidence intervals.
Currently, it supports \code{"perc"},
percentile bootstrap confidence
intervals, and \code{"bc"}, bias-corrected
bootstrap confidence interval.}

\item{...}{Optional arguments.
Ignored.}
}
\value{
A \emph{p} by 2 matrix of the confidence
intervals, \emph{p} being the number
of coefficients.
}
\description{
Return the confidence
interval of the regression
coefficients in the output of
\code{\link[=lm_betaselect]{lm_betaselect()}}.
}
\details{
The type of
confidence intervals depends
on the object. If bootstrapping
was requested, by default it returns
the percentile bootstrap confidence
intervals. Otherwise, it returns the
OLS (or WLS) confidence intervals
and raises a warning for the
standardized solution.

Support for other type of
confidence intervals will be
added.
}
\examples{

data(data_test_mod_cat)

# bootstrap should be set to 2000 or 5000 in real studies
lm_beta_x <- lm_betaselect(dv ~ iv*mod + cov1 + cat1,
                           data = data_test_mod_cat,
                           to_standardize = "iv",
                           do_boot = TRUE,
                           bootstrap = 100,
                           iseed = 1234)
confint(lm_beta_x)
confint(lm_beta_x, method = "ls")
confint(lm_beta_x, type = "raw")

}
\seealso{
\code{\link[=lm_betaselect]{lm_betaselect()}}
}
\author{
Shu Fai Cheung \url{https://orcid.org/0000-0002-9871-9448}
}
