% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print_std_selected_lavaan.R
\name{print.lav_betaselect}
\alias{print.lav_betaselect}
\title{Print a 'lav_betaselect' Object}
\usage{
\method{print}{lav_betaselect}(x, ..., nd = 3, output = c("table", "text"), standardized_only = FALSE)
}
\arguments{
\item{x}{Object of the class
\code{std_solution_boot}.}

\item{...}{Optional arguments to be
passed to \code{\link[=print]{print()}} methods.}

\item{nd}{The number of digits
after the decimal place. Default
is 3.}

\item{output}{String. How the results
are printed. Default is \code{"table"} and
the results are printed in a table
format similar to that of
\code{\link[lavaan:parameterEstimates]{lavaan::parameterEstimates()}}
with \code{output} set to \code{"table"}.
If \code{"text"}, the results will be
printed in a text format similar to
the printout of the output of
\code{\link[=summary]{summary()}} of
a \linkS4class{lavaan} object.}

\item{standardized_only}{Logical.
If \code{TRUE}, only the
results for the standardized solution
will be printed. If \code{FALSE},
the default, then
the standardized solution is printed
alongside the unstandardized solution,
as in the printout of the output
of \code{\link[=summary]{summary()}} of a \linkS4class{lavaan}
object.}
}
\value{
\code{x} is returned invisibly. Called for its side effect.
}
\description{
Print method for an
'lav_betaselect' object, which
is the output of
\code{\link[=lav_betaselect]{lav_betaselect()}}.
}
\details{
The default format of the printout
is that of \code{\link[lavaan:parameterEstimates]{lavaan::parameterEstimates()}},
which is compact but not easy to
read. Users can request a format
similar to that of the printout
of the summary of a \code{lavaan} output
by setting \code{output} to \code{"text"}.

For the \code{"text"} format, users can
also select whether
only the standardized solution is
printed or whether
the standardized solution is appended
to the right of the printout, as
in the output of \code{\link[=summary]{summary()}}
for a \code{lavaan} output.
}
\examples{
library(lavaan)
mod <-
"
med ~ iv + mod + iv:mod
dv ~ med + iv
"
fit <- sem(mod,
           data_test_medmod,
           fixed.x = TRUE)
summary(fit)
fit_beta <- lav_betaselect(fit,
                           to_standardize = c("iv", "dv"))
fit_beta
print(fit_beta,
      output = "text")

}
\seealso{
\code{\link[=lav_betaselect]{lav_betaselect()}}. This
function is adapted from
\code{\link[semhelpinghands:print.std_solution_boot]{semhelpinghands::print.std_solution_boot()}}.
}
\author{
Shu Fai Cheung
\url{https://orcid.org/0000-0002-9871-9448}
}
